requirements.txt
Flask
python-dotenv
opencv-python
yt-dlp
transformers
torch
Pillow